MCP/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
websockets>=11.0.3
requests>=2.28.2
python-dotenv>=1.0.0
aiohttp>=3.8.4
playwright>=1.40.0
pydantic>=2.0.0
jinja2>=3.1.2
markdown>=3.4.3
boto3>=1.28.0
gitpython>=3.1.31
pillow>=10.0.0